crypto: bad `prng_randombytes()` proto (thanks, id0!)
FossilOrigin-Name: d115e90d6dc03b913170f7d4bc821f1a90628a206f5898073153162eb64d6f1b

+++ libs/core/crypto/prng_randombytes.c
@@ -381,7 +381,7 @@ int prng_is_strong_seed () {
 //  prng_randombytes
 //
 //==========================================================================
-void prng_randombytes (void *p, size_t len) {
+void prng_randombytes (void *p, unsigned len) {
   if (!len || !p) return;
 
   // spinlock
